requirements.txt
torch
torchvision
gym==0.21.0
atari_py
opencv-python
numpy
psutil
pyprind
-e git+https://github.com/astooke/rlpyt.git#egg=rlpyt
tqdm
tensorboard
moviepy
# shutil
# pickle
pillow
scikit-learn
matplotlib
lpips
scikit-image
ruamel.yaml
xmltodict